# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "Accion": {
        "enum": [
          "ejecutar",
          "editar-insumos",
          "congelar",
          "recalcular"
        ],
        "title": "Accion",
        "type": "string"
      },
...
      "DetalleCaso": {
        "properties": {
          "acciones_disponibles": {
            "items": {
              "$ref": "#/components/schemas/Accion"
            },
            "title": "Acciones Disponibles",
            "type": "array"
          },
          "actualizado_en": {
            "format": "date-time",
            "title": "Actualizado En",
            "type": "string"
          },
          "actualizado_por": {
            "title": "Actualizado Por",
            "type": "string"
          },
          "descripcion": {
            "title": "Descripcion",
            "type": "string"
          },
          "estado": {
            "$ref": "#/components/schemas/Estado"
          },
          "id_caso": {
            "title": "Id Caso",
            "type": "string"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          },
          "terna": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Terna"
              },
              {
                "type": "null"
              }
            ],
            "description": "Presente solo si el caso tiene una corrida"
          },
          "tipo": {
            "enum": [
              "sin-proyecto",
              "monometalico",
              "polimetalico"
            ],
            "title": "Tipo",
            "type": "string"
          }
        },
        "required": [
          "id_caso",
          "nombre",
          "tipo",
          "estado",
          "actualizado_en",
          "actualizado_por",
          "descripcion",
          "acciones_disponibles"
        ],
        "title": "DetalleCaso",
        "type": "object"
      },
...
      "Estado": {
        "enum": [
          "borrador",
          "calculada",
          "congelada"
        ],
        "title": "Estado",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NuevoCaso": {
        "properties": {
          "descripcion": {
            "default": "",
            "maxLength": 1000,
            "title": "Descripcion",
            "type": "string"
          },
          "duplicar_de": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Identificador del caso a duplicar. El duplicado nace en borrador",
            "title": "Duplicar De"
          },
          "nombre": {
            "maxLength": 120,
            "minLength": 3,
            "title": "Nombre",
            "type": "string"
          },
          "tipo": {
            "enum": [
              "sin-proyecto",
              "monometalico",
              "polimetalico"
            ],
            "title": "Tipo",
            "type": "string"
          }
        },
        "required": [
          "nombre",
          "tipo"
        ],
        "title": "NuevoCaso",
        "type": "object"
      },
...
      "Problema": {
        "description": "Cuerpo de error. Explica qué pasó y cómo resolverlo.",
        "properties": {
          "detalle": {
            "title": "Detalle",
            "type": "string"
          },
          "restriccion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Restricción del cuadro de control que bloquea, si aplica",
            "title": "Restriccion"
          }
        },
        "required": [
          "detalle"
        ],
        "title": "Problema",
        "type": "object"
      },
...
      "ResumenCaso": {
        "properties": {
          "actualizado_en": {
            "format": "date-time",
            "title": "Actualizado En",
            "type": "string"
          },
          "actualizado_por": {
            "title": "Actualizado Por",
            "type": "string"
          },
          "estado": {
            "$ref": "#/components/schemas/Estado"
          },
          "id_caso": {
            "title": "Id Caso",
            "type": "string"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          },
          "tipo": {
            "enum": [
              "sin-proyecto",
              "monometalico",
              "polimetalico"
            ],
            "title": "Tipo",
            "type": "string"
          }
        },
        "required": [
          "id_caso",
          "nombre",
          "tipo",
          "estado",
          "actualizado_en",
          "actualizado_por"
        ],
        "title": "ResumenCaso",
        "type": "object"
      },
...
      "Terna": {
        "description": "Las tres versiones que definen una corrida.",
        "properties": {
          "datos_maestros": {
            "description": "Comité de Precios utilizado",
            "examples": [
              "CP-2026-03"
            ],
            "title": "Datos Maestros",
            "type": "string"
          },
          "huella_inputs": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Huella Inputs",
            "type": "string"
          },
          "motor": {
            "description": "Versión del motor de cálculo",
            "examples": [
              "1.2.0"
            ],
            "title": "Motor",
            "type": "string"
          },
          "revision_inputs": {
            "minimum": 1.0,
            "title": "Revision Inputs",
            "type": "integer"
          },
          "vigencia_datos_maestros": {
            "format": "date",
            "title": "Vigencia Datos Maestros",
            "type": "string"
          }
        },
        "required": [
          "motor",
          "datos_maestros",
          "vigencia_datos_maestros",
          "revision_inputs",
          "huella_inputs"
        ],
        "title": "Terna",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "name": "INVA · Servicio INVA-01-2026-182"
    },
    "description": "\nServicios de la Plataforma de Evaluación Económica · Modelo MINSUR.\n\nReproduce la lógica del modelo económico corporativo para que el área de\nProyectos evalúe escenarios de forma autónoma, con trazabilidad completa.\n\n**Autenticación.** Token de Microsoft Entra ID del tenant corporativo. El\nperfil se deriva de la reclamación `groups`; la plataforma no consulta la\nmembresía a Microsoft Graph.\n\n**Archivos.** Las plantillas se cargan y los reportes se descargan\ndirectamente contra el almacenamiento, mediante firmas de acceso compartido\nde corta vigencia. No atraviesan esta API.\n\n**Respuestas 501.** Indican una operación cuya implementación depende de un\ninsumo del cliente aún no recibido. El cuerpo incluye la restricción del\ncuadro de control que la bloquea.\n",
    "title": "Plataforma de Evaluación Económica · MINSUR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/casos": {
      "get": {
        "operationId": "listar_api_casos_get",
        "parameters": [
          {
            "description": "Filtra por tipo de caso",
            "in": "query",
            "name": "tipo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtra por tipo de caso",
              "title": "Tipo"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResumenCaso"
                  },
                  "title": "Response Listar Api Casos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listar casos",
        "tags": [
          "Casos"
        ]
      },
      "post": {
        "description": "Crea un caso en estado borrador.\n\nCon `duplicar_de`, copia los insumos de un caso existente. El duplicado\nnace en borrador y con revision de inputs 1: es un caso nuevo, no una\nversion del anterior.",
        "operationId": "crear_api_casos_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NuevoCaso"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleCaso"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Crear o duplicar un caso",
        "tags": [
          "Casos"
        ]
      }
    },
    "/api/casos/{id_caso}": {
      "get": {
        "description": "Incluye las acciones disponibles para el perfil del solicitante.",
        "operationId": "detalle_api_casos__id_caso__get",
        "parameters": [
          {
            "in": "path",
            "name": "id_caso",
            "required": true,
            "schema": {
              "examples": [
                "CASO-SR-2026-014"
              ],
              "title": "Id Caso",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleCaso"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detalle de un caso",
        "tags": [
          "Casos"
        ]
      }
    },
    "/api/casos/{id_caso}/comparar/{id_otro}": {
      "get": {
        "description": "Compara el caso sin proyecto contra uno o varios casos con proyecto.\n\nAdvierte cuando las corridas no son comparables: contrastar un caso base\ncalculado con una version mayor del motor contra otro calculado con una\nversion mayor distinta produce una diferencia que no es atribuible al\nproyecto.",
        "operationId": "comparar_api_casos__id_caso__comparar__id_otro__get",
        "parameters": [
          {
            "in": "path",
            "name": "id_caso",
            "required": true,
            "schema": {
              "title": "Id Caso",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id_otro",
            "required": true,
            "schema": {
              "title": "Id Otro",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Comparar Api Casos  Id Caso  Comparar  Id Otro  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Comparar dos casos",
        "tags": [
          "Casos"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "A través de API Management",
      "url": "/"
    }
  ]
}